State changes, slideOver rendering, auth bug fix
