+++ client/src/features/Workspace.jsx
@@ -2,14 +2,17 @@ import Header from '../components/Header';
 import WorkspaceHeader from './WorkspaceHeader';
 import KanbanBoard from './kanbanBoard/KanbanBoard';
 import {useSelectedBoardContext} from '../contexts/BoardContext/boardContext';
-import WorkspaceSlideOver from './WorkspaceSlideOver';
-import { useModalContext } from '../contexts/ModalContext/ModalContext';
+import WorkspaceSlideOver from './WorkspaceSlideover/WorkspaceSlideOver';
+import {useModalContext} from '../contexts/ModalContext/ModalContext';
 
 const Workspace = () => {
-    const {selectedBoard, selectedTask, selectedColumn} = useSelectedBoardContext();
+    const {selectedBoard, selectedTaskId, selectedColumn} =
+        useSelectedBoardContext();
     const {isSlideOverOpen, setIsSlideOverOpen} = useModalContext();
 
     const {...boardInfo} = selectedBoard;
+    console.log(selectedBoard);
+    console.log(boardInfo);
 
     // console.log(boardInfo);
     return (
@@ -17,8 +20,6 @@ const Workspace = () => {
             <Header />
             <WorkspaceHeader boardInfo={boardInfo} />
             <KanbanBoard boardInfo={boardInfo} />
-            {isSlideOverOpen&&<WorkspaceSlideOver  boardInfo={boardInfo} taskInfo={selectedTask} columnInfo={selectedColumn}/>}
-            {/* <WorkspaceSlideOver  boardInfo={boardInfo} taskInfo={selectedTask}/> */}
         </div>
     );
 };
