+++ server/controllers/tasksController.js
@@ -9,10 +9,10 @@ module.exports = {
             // find the specific board
             const board = await Board.findById(boardId);
             // find the specific column
-            const column = board.columns.id(columnId)
-            
+            const column = board.columns.id(columnId);
+
             if (column) {
-                console.log(column)
+                console.log(column);
                 const {tasks} = column;
                 return res.json(tasks);
             }
@@ -35,7 +35,7 @@ module.exports = {
                 taskDetail,
             };
             const board = await Board.findById(boardId);
-            const column = board.columns.id(columnId)
+            const column = board.columns.id(columnId);
             if (!column) {
                 res.json('Column not found');
             }
@@ -58,29 +58,32 @@ module.exports = {
             // const board = await Board.findById(boardId)
             console.log(boardId);
 
-            const updatedTask = {
-                // may need to use spread operator to get all of the task properties
-                taskName,
-                priority,
-                taskDetail
-            };
-
             const board = await Board.findById(boardId);
-            const column = board.columns.id(columnId)
 
             if (!board) {
-                res.json('Board not found');
+                return res.status(404).json({error: 'Board not found'});
             }
+
+            const column = board.columns.id(columnId);
+
             if (!column) {
-                res.json('Column not found');
+                return res.status(404).json({error: 'Column not found'});
             }
 
             const task = column.tasks.id(taskId);
-            task.set(updatedTask);
+
+            if (!task) {
+                return res.status(404).json({error: 'Task not found'});
+            }
+
+            // Update Task properties
+            task.taskName = taskName;
+            task.priority = priority;
+            task.taskDetail = taskDetail;
             await board.save();
 
-            console.log(board);
-            return res.json(board);
+            console.log(task);
+            return res.json(task);
         } catch (error) {
             console.error(error);
             return res.status(500);
@@ -90,10 +93,10 @@ module.exports = {
 
     deleteTask: async (req, res) => {
         try {
-            const {boardId, columnId,taskId} = req.params;
+            const {boardId, columnId, taskId} = req.params;
 
             const board = await Board.findById(boardId);
-            const column = board.columns.id(columnId)
+            const column = board.columns.id(columnId);
             column.tasks.id(taskId).deleteOne();
             await board.save();
             console.log(board);
